Add a property to Field nodes
This property is a boolean that tells if the key was written between
brackets. The utility is when you need to tell the difference between
these two cases: t = {key = 5, [key] = 2}, where t could be indexed with
a string (t.key or t['key']) but also with whatever the value of the
variable key could be.